// src/app/_dto/aiinvoice-detail.dto.ts
export class AiInvoicedetailDto {

    coCode: string;

    dvCode: string;

    brCode: string;

    intCode: string;

    vrNo: number;

    srNo: number;

    vrDate: Date | null;

    itemCode: string | null;

    itemQty: number;

    cFactor: number;

    baseRate: number | null;

    itemRate: number;

    costRate: number;

    addAmt: number;

    curCode: string | null;

    curRate: number;

    disPer: number;

    disAmt: number;

    signInt: number | null;

    uom: string | null;

    interBrCode: string | null;

    storeCode: string | null;

    refCo: string | null;

    refDv: string | null;

    refBr: string | null;

    refInt: string | null;

    refVrNo: number | null;

    refSrNo: number | null;

    retQty: number;

    itemDesc: string | null;

    sortOrder: number | null;

    void: string | null;

    sysuser1: string | null;

    sysdate1: Date | null;

    modiuser: string | null;

    modidate: Date | null;

    jobCode: string | null;

    jobHead: string | null;

    groupCode: string | null;

    subGroupCode: string | null;

    expHead: string | null;

    remark: string | null;

    grnDnRetQty: number;

    vatPer: number;

    vatAmt: number;

    vatType: string | null;

    vat: string | null;

    vatInout: string | null;

    vatCat: number | null;

    vatAc: string | null;

    vatRcmCr: string | null;

    selected?:boolean;

}